+++ src/app/components/team-planner/team-planner.component.scss
@@ -116,6 +116,7 @@ mat-card-content {
     margin-bottom: 10px;
     max-width: calc(6.66% - 10px);
     min-width: 90px;
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
 }
 
 .search-bar {
@@ -186,4 +187,28 @@ mat-card-content {
         flex: 0 0 calc(33.33% - 10px);
         max-width: calc(33.33% - 10px);
     }
+}
+
+.cdk-drag-preview {
+    box-sizing: border-box;
+    border-radius: 4px;
+    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
+                0 8px 10px 1px rgba(0, 0, 0, 0.14),
+                0 3px 14px 2px rgba(0, 0, 0, 0.12);
+}
+
+.cdk-drag-placeholder {
+    opacity: 0;
+}
+
+.cdk-drag-animating {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.character-grid.cdk-drop-list-dragging .character-item:not(.cdk-drag-placeholder) {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+}
+
+.character-item {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
 }
