+++ src/app/models/team.model.ts
@@ -1,0 +1,12 @@
+import { Unit } from "./user-data/unit.model";
+
+export interface Team {
+    name: string;
+    units: Unit[];
+  }
+  
+  export interface Category {
+    name: string;
+    teams: Team[];
+  }
+
