+++ src/app/services/unit/unit-service.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { UnitService } from './unit.service';
+
+describe('Unitservice', () => {
+  let service: UnitService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(UnitService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
